chore(api): rename to dispatchers
