+++ packages/botonic-api/package.json
@@ -43,5 +43,6 @@
     "utf-8-validate": "^5.0.5",
     "uuid": "^8.3.2",
     "ws": "^7.4.6"
-  }
+  },
+  "sideEffects": false
 }
